+++ src/main/java/net/countercraft/movecraft/warfare/MovecraftWarfare.java
@@ -7,11 +7,11 @@ import net.countercraft.movecraft.warfare.bar.config.PlayerManager;
 import net.countercraft.movecraft.warfare.config.Config;
 import net.countercraft.movecraft.warfare.features.assault.AssaultManager;
 import net.countercraft.movecraft.warfare.features.assault.RegionDamagedSign;
-import net.countercraft.movecraft.warfare.features.assault.commands.AssaultCommand;
-import net.countercraft.movecraft.warfare.features.assault.commands.AssaultInfoCommand;
-import net.countercraft.movecraft.warfare.features.assault.commands.AssaultRepairCommand;
+import net.countercraft.movecraft.warfare.commands.AssaultCommand;
+import net.countercraft.movecraft.warfare.commands.AssaultInfoCommand;
+import net.countercraft.movecraft.warfare.commands.AssaultRepairCommand;
 import net.countercraft.movecraft.warfare.features.assault.listener.AssaultExplosionListener;
-import net.countercraft.movecraft.warfare.features.siege.SiegeCommand;
+import net.countercraft.movecraft.warfare.commands.SiegeCommand;
 import net.countercraft.movecraft.warfare.features.siege.SiegeLeaderListener;
 import net.countercraft.movecraft.warfare.features.siege.SiegeManager;
 import net.countercraft.movecraft.warfare.localisation.I18nSupport;
