+++ src/main/java/net/countercraft/movecraft/warfare/commands/SiegeCommand.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.warfare.features.siege;
+package net.countercraft.movecraft.warfare.commands;
 
 import net.countercraft.movecraft.craft.Craft;
 import net.countercraft.movecraft.craft.CraftManager;
@@ -7,6 +7,9 @@ import net.countercraft.movecraft.repair.MovecraftRepair;
 import net.countercraft.movecraft.util.TopicPaginator;
 import net.countercraft.movecraft.warfare.MovecraftWarfare;
 import net.countercraft.movecraft.warfare.config.Config;
+import net.countercraft.movecraft.warfare.features.siege.Siege;
+import net.countercraft.movecraft.warfare.features.siege.SiegeManager;
+import net.countercraft.movecraft.warfare.features.siege.SiegeUtils;
 import net.countercraft.movecraft.warfare.features.siege.events.SiegeBroadcastEvent;
 import net.countercraft.movecraft.warfare.features.siege.events.SiegePreStartEvent;
 import net.countercraft.movecraft.warfare.localisation.I18nSupport;
